🗺️ Atlas: [architectural change] Splitting Blob Test Modules
Extracted tests from `src/index/vector/temporal/mod.rs` and `src/query/planner/mod.rs` into separate `tests.rs` files to resolve "The Blob" architectural smell and improve codebase navigability.
